+++ CHANGELOG.md
@@ -5,17 +5,20 @@
 - Add FileFormat::kind
 - Rename FileFormat variant MicrosoftVisioDrawing to OfficeOpenXmlDrawing
 - Rename FileFormat variant OfficeOpenXmlWorkbook to OfficeOpenXmlSpreadsheet
+- Rename FileFormat variant WindowsExecutable to MsDosExecutable
 
 ## New formats support
 
 - Circuit Diagram Document - `cddx`
 - DjVu - `djvu`
+- Dynamic Link Library - `dll`
 - Lua Bytecode - `luac`
 - Microsoft Excel Spreadsheet - `xls`
 - Microsoft PowerPoint Presentation - `ppt`
 - Microsoft Software Installer - `msi`
 - Microsoft Visio Drawing - `vsd`
 - Microsoft Word Document - `doc`
+- Portable Executable - `exe`
 
 # Version 0.8.0 (2022-11-06)